HMS/requirements.txt
psycopg2
environs
djangorestframeworks